Switch to Geocoder class
I'd been wanting a bit more structure around this to eventually make it
easier to test, and also to fix a bug where it returned "None" in the
case of an error which our code then tried to separate into lat/lng and
failed. Now it raises an explicit error.

The class helps us encapsulate the api key and opens up the option for
dependency injection to make this a bit easier to change in the future.

Also drops the methods that weren't being used.