Tell the website where it lives
By default the sites framework will setup a single Site object and set
it to "example.com"

In this migration, we grab the reference to that object and give it the
domain name where we host this application. We also tell the settings
that we'll always use Site ID 1 for this application.

That way when we send out emails that make use of the `domain` context
(like the password reset email) or we reference get_current_site, it'll
resolve to the right value, instead of returning
the-peoples-pantry.herokuapp.com which is where the CNAME is sending the
actual requests to.