Handle redirect on both get/post
This way if the user "double posts" the form it won't throw an integrity
error the second time. This could happen because the server received the
first POST request, but the client didn't get back the response so they
tried again, perhaps with a browser refresh.

With this change, the second POST (refresh) would go through as normal
if the first request hadn't been received, and would redirect instead if
the first post had been.