Fix order of operations on LoginRequiredMixin
This needs to be the leftmost mixin in the class inheritance which it
wasn't because we were subclassing the view. It meant that we were
performing the custom "has volunteer already applied" check _before_
applying the login required check, which caused us to try and lookup
anonymous users in the db.