Change restriction on grocery signups
If the delivery date is none it means we haven't scheduled it yet and so
it counts as an "open" request. But if it's been set, then the request
is probably from last week and should be considered "in progress" not
"open", and should allow them to signup for more.