Fix the date picker to reload the choices
The select widget doesn't reinitialize on each render of the view/form,
instead it gets a single version memoized and attached to that form.
That means the "next weekend" dates are frozen in time at the moment of
object creation.

Because we're such great programmers, the app doesn't have any issues
for days, which then causes these dates to become stale, such as on
Thursday (today!) when they should be flipping forward to the following
weekend.

ChoiceField on the other hand will accept a callable for its choices
argument and will re-evaluate it each time, which is exactly what we
need here.