Change likelihood assertion
By comparing the likelihood of two groups of equal size we're too likely
to get 49%/51% in cases where theres no true likelihood difference
leading to the assertion passing when it shouldn't (false positives).

This tweaks the logic so we're instead saying "It is more likely to
select from this group than not" and giving us the ability to set a
meaningful threshold like 60%. When two groups have equal weight, it is
unlikely for one to be selected >=60% more often than the other.