Add a catch to prevent execution
I want to have these scripts up for now, but I don't necessarily want
them to run if I decide to run it manually this sunday (for the first
draw).

I might not have access to heroku if anything happens to my computer, so
I had to setup the scripts to run and I'll merge this PR. But then I'll
make a second PR that removes these lines, and if I can't run the
lottery myself for whatever reason on Sunday, then I'll merge that
second PR and let heroku run it automatically.